requirements.txt
asyncio==4.0.0
bm3d==4.0.3
bm4d==4.2.5
build==1.3.0
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
filelock==3.19.1
fonttools==4.60.0
fsspec==2025.9.0
imageio==2.37.0
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
lightning-utilities==0.15.2
lpips==0.1.4
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pyparsing==3.2.5
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.9.0
scikit-image==0.25.2
scipy==1.16.2
seaborn==0.13.2
setuptools==75.1.0
six==1.17.0
sympy==1.14.0
tifffile==2025.9.20
torch==2.8.0
torchmetrics==1.8.2
torchvision==0.23.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
wheel==0.44.0
